+++ examples/main.rs
@@ -10,12 +10,12 @@ fn main() -> BinaryMerkleTreeResult<()> {
 
     // Inserting and getting from a tree
     let new_root = tree.insert(None, &mut [&key], &mut vec![&value])?;
-    let items = tree.get(&new_root, &mut [&key])?;
-    assert_eq!(items, vec![Some(value)]);
+    let item_map = tree.get(&new_root, &mut [&key])?;
+    assert_eq!(item_map[&key[..]], Some(value));
 
     // Attempting to get from a removed root will yield None
     tree.remove(&new_root)?;
-    let items2 = tree.get(&new_root, &mut [&key])?;
-    assert_eq!(items2, vec![None]);
+    let item_map2 = tree.get(&new_root, &mut [&key])?;
+    assert_eq!(item_map2[&key[..]], None);
     Ok(())
 }